Add PlatformEncoding and Utf8PlatformEncoding
Adds new `PlatformEncoding` and `Utf8PlatformEncoding` structs that implement the same logic as `NativeEncoding` and `Utf8NativeEncoding` but are structs instead of type aliases.

This enables stronger type enforcement without the need for conditional config as `PlatformEncoding` is not the same as `UnixEncoding` or `WindowsEncoding`, regardless of how it is implemented at compile-time.

### Tests

Run `cargo test platform --doc`.